+++ c3DSkit/c3DSkit.c
@@ -3,14 +3,21 @@
 #include <numpy/arrayobject.h>
 #include <stdlib.h>
 #include <stdint.h>
-#include <time.h>
+#include <string.h>
 
 #define bool uint8_t
 #define true 1
 #define false 0
 
+#define ABS(val) ((val < 0) ? -val : val)
+#define LOG2(val) (log(val) / log(2))
+#define MIN(a,b) (((a)<(b))?(a):(b))
+#define MAX(a,b) (((a)>(b))?(a):(b))
+#define CLAMP(val, low, high) ((val < low) ? low : ((val > high) ? high : val))
+
 #include "audio.h"
 #include "compression.h"
+#include "graphics.h"
 
 static PyMethodDef functions[] = {
 	{"decodeDSPADPCMblock", decodeDSPADPCMblock, METH_VARARGS, "Decodes a DSPADPCM sample block into 16-bits PCM data"},
@@ -18,6 +25,8 @@ static PyMethodDef functions[] = {
 	{"generateDSPADPCMcoefs", generateDSPADPCMcoefs, METH_VARARGS, "Generate coefficients for DSPADPCM encoding"},
 	{"compressLZ11", compressLZ11, METH_VARARGS, "Compress a byte stream in LZ11"},
 	{"decompressLZ11", decompressLZ11, METH_VARARGS, "Decompress a LZ11 compressed stream"},
+	{"extractTiledTexture", extractTiledTexture, METH_VARARGS, "Extracts a tiled texture (like BFLIM ones) to an RGBA byte array"},
+	{"getTextureFormatId", getTextureFormatId, METH_VARARGS, "Returns the internal format ID from its name"},
 	{NULL, NULL, 0, NULL}
 };
 
@@ -29,6 +38,5 @@ static struct PyModuleDef c3DSkitdef = {
 
 PyMODINIT_FUNC PyInit_c3DSkit(void){
 	import_array();
-	srand(time(NULL));
 	return PyModule_Create(&c3DSkitdef);
 }